separate db into their own files
