+++ controllers/friendsController.js
@@ -24,7 +24,7 @@ module.exports = {
     update: function(req, res) {
         db.Friend   
             .findOneAndUpdate({ _id: req.params.id }, body)
-            .then(dbModel => res.json(dbMoel))
+            .then(dbModel => res.json(dbModel))
             .catch(err => res.status(422).json(err))
     },
     delete: function(req, res) {
