+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "express": "^4.16.3",
+    "fs": "0.0.1-security",
     "if-env": "^1.0.4",
     "mongoose": "^5.5.2",
     "react-router-dom": "^5.0.0"
